Move employee related admin.py code to employee app
